wip: split marketing performance adapters
